api-server/requirements.txt
Flask
openai
Werkzeug
python-dotenv
flask-cors
python-chess
chromadb
gunicorn